Extensive refactoring to allow for more flexible percolation rules
Percolation rules can now be specified as "site" or "bond" rules that
are either describing rules for when a site of a given sublattice is
stable (site rules) or for when the bond between two neighboring sites
of possibly two different sublattices is percolating (bond rules).

The code should now be fully Python 3 complient (where it is not, it is
a bug).  Python 2.7 compatibility should be given but is untested.